Location resolution finished.

+++ src/CodeGenerationMessage.cpp
@@ -7,7 +7,7 @@
 
 CodeGenerationMessage::CodeGenerationMessage(const bool &is_offset,
                                              const bool &is_register,
-                                             const int &value) {
+                                             const long long int &value) {
     this->is_offset = is_offset;
     this->is_register = is_register;
     this->value = value;
